+++ app/lib/modules/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.server';
+export * from './session.server';
