+++ app/lib/modules/money/index.ts
@@ -1,0 +1,1 @@
+export * from './money-utils';
