+++ app/lib/modules/stack-categories/components/categorized-stacks.tsx
@@ -4,8 +4,7 @@ import { DragDropContext, DropResult } from 'react-beautiful-dnd';
 import { useFetcher } from '@remix-run/react';
 import { DraggableItem } from '~/components/beautiful-dnd-wrappers/draggable-item';
 import { DroppableList } from '~/components/beautiful-dnd-wrappers/droppable-list';
-import { EditableStack } from '~/lib/modules/stacks/components/editable-stack';
-import { recalcStackPositions } from '~/lib/modules/stacks/utils/stack';
+import { EditableStack, recalcStackPositions } from '~/lib/modules/stacks';
 
 export type CategoryWithStack = Prisma.StackCategoryGetPayload<{ include: { Stack: true } }>;
 
@@ -117,4 +116,4 @@ function CategorizedStacks({ categorized }: { categorized: CategoryWithStack[] }
   );
 }
 
-export default CategorizedStacks;
+export { CategorizedStacks };
