+++ app/lib/modules/stack-categories/index.ts
@@ -1,0 +1,3 @@
+export * from './stack-category-utils';
+export * from './stack-category-utils.server';
+export * from './components/categorized-stacks';
