+++ app/lib/modules/stack-categories/stack-category-utils.server.ts
