+++ app/routes/budget/stack-category/$stackCategoryId.tsx
@@ -3,8 +3,8 @@ import { Form, useLoaderData } from '@remix-run/react';
 import { StackCategory } from '.prisma/client';
 import { db } from '~/lib/db.server';
 import { Button } from '~/components/button';
-import { requireAuthenticatedUser } from '~/lib/modules/user/utils/user.server';
-import { deleteStackCateogry } from '~/lib/modules/stack-categories/utils/stack-category.server';
+import { requireAuthenticatedUser } from '~/lib/modules/user';
+import { deleteStackCateogry } from '~/lib/modules/stack-categories';
 
 interface LoaderData {
   category: StackCategory;
