+++ app/routes/login.tsx
@@ -1,5 +1,5 @@
 import { LoaderFunction, redirect } from '@remix-run/node';
-import { authenticator } from '~/lib/modules/auth/auth.server';
+import { authenticator } from '~/lib/modules/auth';
 
 export const loader: LoaderFunction = async ({ request }) => {
   const user = await authenticator.isAuthenticated(request);
